+++ tests/integration/models/reddit/test_subreddit.py
@@ -34,7 +34,6 @@ from praw.models import (
     ModmailConversation,
     ModmailMessage,
     Redditor,
-    Stylesheet,
     Submission,
     Subreddit,
     SubredditMessage,
@@ -2018,12 +2017,12 @@ class TestSubredditStylesheet(IntegrationTest):
     def subreddit(self):
         return self.reddit.subreddit(pytest.placeholders.test_subreddit)
 
-    def test_call(self):
-        with self.recorder.use_cassette("TestSubredditStylesheet.test_call"):
-            stylesheet = self.subreddit.stylesheet()
-        assert isinstance(stylesheet, Stylesheet)
-        assert len(stylesheet.images) > 0
-        assert stylesheet.stylesheet != ""
+    def test_attrs(self):
+        self.reddit.read_only = False
+        stylesheet = self.subreddit.stylesheet
+        with self.recorder.use_cassette("TestSubredditStylesheet.test_attrs"):
+            assert len(stylesheet.images) > 0
+            assert stylesheet.stylesheet != ""
 
     def test_delete_banner(self):
         self.reddit.read_only = False